requirements.txt
flask==3.1.0
pdfplumber==0.11.4
python-dotenv==1.1.0
gunicorn==23.0.0
